add rest-plant-api learning post stub
For some reason, forestry wont let you create a frontmatter template
then a post immediatly based on it. So, instead, on forestry I've
created the template, then here I've created a content file with nothing
in it. Hopefully, when pushed, forestry will then let me add the
frontmatter fields to it and create more of them.